Move actor to job package
