+++ gitmesh-repository-server/src/main/scala/com/github/takezoe/gitmesh/repository/api/models.scala
@@ -2,4 +2,10 @@ package com.github.takezoe.gitmesh.repository.api
 
 object models {
 
+  case class CloneRequest(nodeUrl: String, empty: Boolean)
+  case class Repository(name: String, empty: Boolean)
+  case class Status(url: String, diskUsage: Double, repos: Seq[String])
+  case class SynchronizedRequest(nodeUrl: String)
+  case class SynchronizeRequest(nodeUrl: String)
+
 }
